Configured the VDF parser to ignore whitespace instead of removing it.
The `RemoteWhitespaces` function in `vdf.cpp` alternates to removing
whitespace chars depending on whether or not its currently in a key or
value. However, it does not check for escaped quotes, and this can cause
the parser to fail to read. Since the parser makes better determinations
of what is and is not a key or a value, the logic to ignore whitespace
was moved directly into the parser itself. The RemoveWhitespace was
taken off the LoadFromText function, since it would be going over the
config.vdf twice.

Additionally, added some troubleshooting steps for Arma 3.

Fixes #284